rename `getAvatar.ts` -> `getEnsAvatar.ts`
